test(egfx): move decoder tests to testsuite, add error coverage
Move the 4 OpenH264 decoder tests from inline #[cfg(test)] in decode.rs
to the testsuite crate per ARCHITECTURE.md testing invariants. Add 4
error-path tests: empty input, truncated NAL length prefix, NAL length
exceeding buffer, and garbage NAL content.

The openh264-bundled feature on the testsuite gates both the test module
and the openh264 source dependency, so regular `cargo test` does not
pull in cc/nasm-rs build deps.
