+++ crates/ironrdp-testsuite-core/tests/egfx/mod.rs
@@ -1,2 +1,4 @@
 mod client;
+#[cfg(feature = "openh264-bundled")]
+mod decode;
 mod server;
